requirements.txt
flask==3.0.0
flask-socketio==5.3.6
pandas==2.1.4
numpy==1.26.2
alpaca-py==0.29.2
python-dotenv==1.0.0
requests==2.31.0
gevent==23.9.1
yfinance==0.2.18
plotly==5.17.0
dash==2.14.1
dash-bootstrap-components==1.5.0
python-crontab==3.0.0
schedule==1.2.0 